Domain whitelist + connector GUI
Added a pageaction available when the api is injected into a page.
From this we can access the whitelist and remove entries if needed, but
it's very basic.

Upon a website calling `ergo_request_read_access()` the connector will
ask the user if they wish to connect if the site is not on the
whitelist, and optionally if they wish to add it to the whitelist. If it
was on the whitelist this confirmation is skipped entirely and the API
is injected right away.

The whitelist is stored in chrome's local storage.